+++ go.mod
@@ -7,4 +7,6 @@ require (
 	github.com/go-sql-driver/mysql v1.6.0
 	github.com/kataras/iris/v12 v12.2.0-beta1
 	github.com/spf13/viper v1.11.0
+	gorm.io/driver/mysql v1.3.3
+	gorm.io/gorm v1.23.5
 )
