+++ src/common/common.go
@@ -4,3 +4,18 @@ type Response struct {
 	StatusCode int32  `json:"status_code"`
 	StatusMsg  string `json:"status_msg,omitempty"`
 }
+type User struct {
+	Id            int64  `json:"id,omitempty"`
+	Name          string `json:"name,omitempty"`
+	FollowCount   int64  `json:"follow_count,omitempty"`
+	FollowerCount int64  `json:"follower_count,omitempty"`
+	IsFollow      bool   `json:"is_follow,omitempty"`
+	Password      string
+}
+
+func (user *User) IsCorrect(password string) bool {
+	return user.Password == password
+}
+func (user *User) Exists() bool {
+	return user.Id > 0
+}
